requirements.txt
requests~=2.32.5
python-dotenv~=1.2.1
SQLAlchemy~=2.0.45
termcolor~=3.1.0
matplotlib~=3.9.4
RapidFuzz~=3.13.0
urllib3~=2.6.2